Be more flexible about Array constructor matching
Since code run in different contexts may have different object-identity
values for the Array constructor, matching on constructor identity can
result in spurious test failures.

To fix this, treat any Array-like (ie, Array.isArray() positive result)
as a "real" Array if its constructor name is "Array".

The reasoning is that if the name of the ctor changes, it probably also
has some added features that make it not strictly the "same" thing.

This is the root cause of https://github.com/tapjs/node-tap/issues/470